added casl ability

+++ controllers/admin/adminsController.js
@@ -2,8 +2,15 @@ import * as adminService from "../../services/admin/adminService.js";
 import asyncHandler from "express-async-handler";
 import ApproveOwnerSchema from "../../zodSchemas/ApproveOwnerSchema.js";
 import ValidateBodyOnSchema from "../../utils/ValidateBodyOnSchema.js";
+import { AppError } from "../../middleware/errorMiddleware.js";
 
 const getOwners = asyncHandler(async (req, res) => {
+  if (req.user.ability.cannot("manage", "Owners")) {
+    throw new AppError({
+      statusCode: 403,
+      message: "No enough permission to read this resource",
+    });
+  }
   const {
     name,
     owner,
@@ -44,6 +51,13 @@ const getOwners = asyncHandler(async (req, res) => {
 });
 
 const deleteOwner = asyncHandler(async (req, res) => {
+  if (req.user.ability.cannot("manage", "Owners")) {
+    throw new AppError({
+      statusCode: 403,
+      message: "No enough permission to perform this action",
+    });
+  }
+
   const { ownerId } = req.params;
   const deletedOwner = await adminService.deleteOwner(ownerId);
   return res.json({
@@ -53,6 +67,12 @@ const deleteOwner = asyncHandler(async (req, res) => {
 });
 
 const approveOwner = asyncHandler(async (req, res) => {
+  if (req.user.ability.cannot("manage", "Owners")) {
+    throw new AppError({
+      statusCode: 403,
+      message: "No enough permission to perform this action",
+    });
+  }
   const data = ValidateBodyOnSchema(req.body, ApproveOwnerSchema);
   const approveOwner = await adminService.approveOwner(
     req.params.ownerId,
@@ -67,11 +87,17 @@ const approveOwner = asyncHandler(async (req, res) => {
 });
 
 const getBalance = asyncHandler(async (req, res) => {
+  if (req.user.ability.cannot("manage", "Transactions")) {
+    throw new AppError({
+      statusCode: 403,
+      message: "No enough permission to read this resource",
+    });
+  }
   const { transactions, balance } = await adminService.getBalance();
 
   return res.json({
     wallet: {
-      balance,
+      balance: balance || 0,
     },
     transactions,
   });
